Wed, Jul 26, 2023, 9:52 PM +01:00

+++ meal/meal.py
@@ -2,8 +2,8 @@ s = input('time: ')
 s = s.split(':')
 print(s)
 
-a = int(s[0])
-b = int(s[1])
+a = float(s[0])
+b = float(s[1])
 
 if 0 < b < 15:
     b = 0
@@ -22,4 +22,4 @@ print(b)
 1 = 60
 0.5 = 30
 x = 20
-x = 40 
+x = 40
